+++ tests/src/Recipe/Cookbook/NewAccountEndPointWithOptionalsTest.php
@@ -36,6 +36,7 @@ use Teknoo\East\Common\Contracts\Recipe\Step\RenderFormInterface;
 use Teknoo\East\Common\Recipe\Step\CreateObject;
 use Teknoo\East\Common\Recipe\Step\RenderError;
 use Teknoo\East\Common\Recipe\Step\SaveObject;
+use Teknoo\East\Paas\Recipe\Step;
 use Teknoo\Recipe\RecipeInterface;
 use Teknoo\Tests\Recipe\Cookbook\BaseCookbookTestTrait;
 
@@ -174,7 +175,19 @@ class NewAccountEndPointWithOptionalsTest extends TestCase
             $this->getRenderError(),
             [
                 static function () {},
-                static function () {},
+                new class {
+                    public function __invoke() {}
+                },
+                [
+                    new class {
+                        public function foo() {}
+                    },
+                    'foo',
+                ],
+                new Step(
+                    static function () {},
+                    ['foo' => 'bar']
+                ),
             ],
             $this->createMock(ObjectAccessControlInterface::class),
             'foo.template',